-changed mouse position in camera from int to float and in the intervall [0,1] (for easier computations) -added selection highlight for arrow heads

+++ src/Visualisation/camera.h
@@ -88,9 +88,9 @@ public:
 	 *  \param x The new position in x direction
 	 *  \param y The new position in y direction
 	 */
-	virtual void set_view_by_mouse(int x, int y);
+	virtual void set_view_by_mouse(float x, float y);
 
-	virtual void set_button_press_mouse(int x, int y);
+	virtual void set_button_press_mouse(float x, float y);
 
 	/**
 	 * \brief This rotates the camera around a point
