+++ src/Visualisation/moveable_camera.h
@@ -56,9 +56,9 @@ public:
 	 * \param x The new mouse position
 	 * \param y The new mouse position
 	 */
-	void set_view_by_mouse(int x, int y);
+	void set_view_by_mouse(float x, float y);
 
-	void set_button_press_mouse(int x, int y);
+	void set_button_press_mouse(float x, float y);
 
 	/**
 	 * \brief This strafes the camera left or right depending on the speed (+/-)
@@ -117,8 +117,8 @@ public:
 	virtual std::string get_name();
 
 private:
-	int down_x_;
-	int down_y_;
+	float down_x_;
+	float down_y_;
 };
